Fix NuGet/Home#213: Restore with WinRT C++ fails
The problem is nuget always uses MSBUILD v4.0 to parse solution files. This will fail with new types of solutions/projects created by VS 2013 & VS 2014. The fix is to add an option -MSBuildVersion to let the user specify the version of msbuild to be used. This option also replaces option -MSBuildPath.