Fixes https://github.com/NuGet/Home/issues/1026.
1. Enabled all the tests which are skipped today and verified that they
pass with the changes mentioned below

2. Uses the ListCommandResource to get the list endpoint. In case of V3
endpoints, it will return the v2 endpoint if configured. That will be used
to list the packages. Added tests for this scenario

3. Old nuget.exe list has a lot of specific behavior
	a. It does not all the versions in case of nuget.org alone
        b. Tries to list all the packages when no search term is provided.
Need to print packages as it queries

Given the above and other unobserved legacy behavior of nuget.exe list,
taking a dependency on NuGet.Core to simply do whatever
the old nuget.exe list does