Fix NuGet/Home#1336: Msbuild selection in nuget.exe 3.2 RC is broken
nuget runs "msbuild.exe /version" to get the version of msbuild in PATH, then looks in registry
to find the installation path of that version. However, the version number of a recent update of msbuild 4.0
is now 4.6.x.x, which does not match any version in the registry. nuget then fails because it cannot
locate the installation path of msbuild.

Instead of trying to use the msbuild in PATH, nuget now just uses the highest installed version.